pyproject.toml
[build-system]
requires = ["setuptools>=75", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openclaw-smart-agent"
version = "0.1.0"
description = "Smart agent runtime, plugin, and skill bundle for OpenClaw."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
  { name = "OpenClaw Smart Agent Contributors" }
]
dependencies = [
  "fastapi>=0.116,<1.0",
  "pydantic>=2.11,<3.0",
  "PyYAML>=6.0,<7.0",
  "uvicorn>=0.35,<1.0"
]

[project.optional-dependencies]
dev = [
  "httpx>=0.28,<1.0",
  "pytest>=8.4,<9.0"
]

[project.scripts]
openclaw-smart-agent = "openclaw_smart_agent.cli:main"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
openclaw_smart_agent = ["templates/*.yaml"]

[tool.pytest.ini_options]
testpaths = ["tests"]
